refactor: testing service instead of repository
